Remove unnecessary passing through of labelContent
